Refactor `Version::parts` into a newtype
Create a type `VersionParts` to house the "parts" and move the custom `Eq`,
`Hash`, and `PartialEq` logic to this struct. This makes it so `Version`
can automatically derive those impls.

This change makes it possible to reuse the custom equality logic for
comparing "parts" and fixes the test failures added in the previous commit.

This overlaps with a similar refactoring in #337 but that PR is not merged
yet.

Signed-off-by: J Robert Ray <jrray@imageworks.com>