Move integration tests to "rust" workflow
Modify the "rust" workflow so that it runs the integration tests, so
building an rpm is not required for every CI run.

Change the "rust" workflow to run in a container, using the "rust" image as
a base, and make tweaks required to allow the workflow to run locally with
nektos/act.

Set the "rpm" workflow to only run on release branches, and increase the
timeout so it can succeed now with the longer build times because of lto.
(I'm assuming that 30 minutes is long enough.)

Signed-off-by: J Robert Ray <jrobertray@gmail.com>