Accept `-l` flag but it does nothing
Print a deprecation warning if the flag is used. This is implemented using
a feature added in Clap 3.2, so bump the version of Clap.

Note that the deprecation warning can be printed multiple times, because
of how spk re-runs itself to activate an spfs runtime. We shouldn't have
to live with this very long.

Closes #358

Signed-off-by: J Robert Ray <jrray@imageworks.com>