Rework `CompatRange` and `WildcardRange` parsing
Parsing version ranges that had versions with multiple tag sets was not
working, such as:

    pkg/1.0+a.0,A.0,2.0+b.0,B.0

Because of the ambiguity with numeric-only tag names, the property testing
was only generating inputs with a single tag in the tag set. But if changed
to generate non-ambiguous cases, it reveals that these types of values
aren't parsed correctly.

Don't assume everything up to the first ',' is the complete version range.

Switch parsing `CompatRange` to nom parsing and require a compatibility
prefix to detect this case, before trying other alternate forms. Change
`wildcard_range` to require a wildcard, so this alternate only matches true
wildcard expressions. Handling empty inputs has moved elsewhere. Happily
this function no longer takes boolean arguments.

Add an alternate that handles the plain version number expression. This is
last so it doesn't consume input like `0.0.*.0.0`. This should be consumed
by `wildcard_range` but would be partially consumed by `compat_range` if
that allowed an optional compatibility prefix.

This is a backport from #370 with some additional changes to make the tests
succeed without the change to require tag names have some non-numeric
character. It now succeeds at parsing but in the situation where an input
has two legal interpretations, it will always prefer to interpret extra
tag set entries rather than a separate range filter.

Signed-off-by: J Robert Ray <jrray@imageworks.com>