Require tag names contain an alpha character
This addresses #329 in a more relaxed sense that it doesn't require the tag
begin with a letter but a letter must appear somewhere in the tag.

With this change, the property test can now generate tag sets with more
than one entry. This ended up revealing a problem in the parser and was
fixed in the previous commit.

The property tests also now generate invalid `Version` instances to verify
those are detected as invalid.

Signed-off-by: J Robert Ray <jrray@imageworks.com>