`BTreeMap` for `State::packages`
Use a map for `State::packages` and cache the hash of the `Spec` part of
the values. A map keyed by `PkgName` is necessary so that when replacing
a solved package with a different build, the old build is removed. This
has the same net effect as appending the new package to the end of a
list, and then processing them in FIFO order. But now the state id will
be the same for two states that have the same solved package in it,
regardless of what other builds of the same package might have been
added previously.

Have to add `Ord` and `PartialOrd` to lots of stuff.

Signed-off-by: J Robert Ray <jrray@imageworks.com>