Remove `CloneableError`
Replace this with `CacheValue` that plays a similar role but doesn't try
to represent every possible error type accurately. The ones that need to
be preserved are represented -- only `PackageNotFoundError` is needed in
testing -- as well as the `String` type which can easily be re-mapped back
onto the same type.

For `LS_TAGS_CACHE`, however, errors in the stream are filtered out and
only successful results are cached.

Signed-off-by: J Robert Ray <jrray@imageworks.com>