WIP: flatten decisions
Working on getting all the decisions for the different builds of a version
of a package and then working on those in parallel.

Difficultly: there is only one `current_node` and a single stream of
(Node, Decision). The stream is just for reporting purposes. `current_node`
is just _some_ position in the solve.