Fix SIGSEGV in `getenv ()` when sentry feature is enabled
There is some kind of data race with another thread that is doing reqwest/
tokio stuff, I believe this is a thread started by sentry but there are no
sentry related stack frames on thread 2. I tried moving up to the latest
release of `whoami` but it didn't help. Grabbing the username before
calling `sentry::init` seems to avoid the race.

Signed-off-by: J Robert Ray <jrray@imageworks.com>