Change history to `PriorityQueue`
The recent change to `VecDeque` didn't quite get the desired behavior. This
use of a priority queue gives history entries that were added higher up in
the solve tree priority. The goal is to exhaust all possible builds of the
top-level package before iterating on builds lower down.

The `VecDeque` version of the code still gets into an endless search for
one known case, and this version will find a solution in 223 seconds.

Signed-off-by: J Robert Ray <jrray@imageworks.com>