flags: allow host options to be overridden
If a package was built on one distro (centos) then we are unable
to activate it using "spk env" on rhel because the requests are
incompatible due to the "distro" field.

One option is to use the --no-host option, but it is a
heavy-hammer that strips out all host options.

A more surgical approach would be to allow single host options
to be overriden.

Currently, "spk env -o distro=centos bash/5.1" does not resolve
when running on rhel because the "distro" option is always
clobbered by the option parser.

Change the order that options are parsed so that user-specified
options on the command-line have override the automatic host
options.

Closes #387
Signed-off-by: David Aguilar <davvid@gmail.com>