Output the solution packages from solvers run by 'make-binary'  (#365)
Adds the solution packages to the output for resolves during 'spk make-binary' commands and 'spk test' commands. This involves moving the solver solution package list output inside the DecisionFormatter, adding '--show-solution' flag, and adding a heading prefix for the 'Installed Packages' heading and `with_header` method to set it. Other commands that output the solution packages are updated to use these changes and use a builder directly when required.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>