Allow `Binary:1.2.3` and `=1.2.3` requests to merge
Normally `Compat` ranges aren't allowed to merge (simplify), but there is
this special case where a `Compat` and `Equals` with matching version
numbers are fair game. The reasoning is that any build that satisfies
`=1.2.3` must also satisfy `Binary:1.2.3`, because version 1.2.3 must be
binary compatible with itself. This is not true of `>=1.2.3`.

We have big solves that contain a variety of cases that match this
situation, as in:

    'python-pyside2:all/=5.15.2,Binary:5.15.2'
    'python:all/=3.7.7,Binary:3.7.7'

Allowing these to merge will help reduce the distinct state id count.

Signed-off-by: J Robert Ray <jrray@jrray.org>