Add `ident_parts`
For `ls --recursive` to parse the command line args, and to know if a
version was specified, `ident_parts` can be used to parse something that
could be an `Ident` but without producing an `Ident` value. An `Ident`
always has a `Version`, so it is not possible to know if the version was
specified by the user.

Some things have to become `pub` to be used from the bin crate.

Signed-off-by: J Robert Ray <jrray@jrray.org>