Handle "local" for `--enable-repo` and `--disable-repo`
With `--local-repo` going away, and some commands like `rm` disabling the
local repo if `--enable-repo origin` is used, then a way to opt back into
enabling the local repo is needed. This change allows the user to use
`--enable-repo local`.

Also, it follows that a user might try to use `--disable-repo local`.

Hide the `--no-local-repo` flag because it becomes redundant and could be
removed.

Signed-off-by: J Robert Ray <jrray@jrray.org>