Add a way to track the owner of an embedded package
Do so by extending `Build::Embedded` to be able to contain the `Ident` of
the package that is embedding that package. This is an enum with an
`Unknown` variant in order to express existing packages that did not have
this information filled in.

This change introduces a new string syntax for embedded builds, where the
`Ident` string is appended to "embedded" inside square brackets:

    embedded-pkg-name/1.2.3/embedded[package-that-embeds/3.4.6/3I42H3S6]

This change causes in a bit of turmoil in the parser code. Because parsing
a `Build` now needs to be able to parse a nested `Ident` the logic for
detecting the end of an `Ident` needs to check for either `eof` or
`']'`. A tagset also might be terminated by a `']'`.

This enum variant is not populated by anything yet, so the tests are
written to expect `Unknown` for now.

Signed-off-by: J Robert Ray <jrray@jrray.org>