Create a `BuildIdent` type
This concept originally appeared in #278, and now has been extended to
incorporate the repository name. There is a need to capture a repository
name and an `Ident` together, but there are a lot of cases where an `Ident`
should not be conceptually bound to a specific repository.

Specifically, when creating a source package in the context of `spk build`,
it is desired to bind the built package to the repo it was published to,
so when building the package the correct source package is resolved.

The `BuildIdent` type sets the stage to be able to remove the repository
name field from `Ident`.

Signed-off-by: J Robert Ray <jrray@jrray.org>