Remove `repository_name` from `Ident`
While working on #415 it became clear that it is confusing to have `Ident`
contain a repository name and when it should be displayed/honored and when
it should be ignored. The reason for adding it was to bind an `Ident` to
a specific repo in some cases, like `BuildSource::SourcePackage`, but in
general an `Ident` transcends what repository it lives in.

In contrast, the repository name on a `RangeIdent` makes sense because a
user may want to name a package in a specific repository. We were missing
some glue to convert between `Version` and `VersionRange`.

Now a `RangeIdent` is used in places where an `Ident` was used before and
the repository name binding is desired.

The ident parsing tests that were put in place to exercise the ambiguous
requests for `Ident` have been retained but use `RangeIdent` instead.

Signed-off-by: J Robert Ray <jrray@jrray.org>