Expand on `DataPath` with `TagPath`
`TagPath` is similar to `DataPath` but produces a path to be used for the
spfs tag name.

An embedded package's spec tag path looks like this:

    spk/spec/embedded-package/1.0/embedded-by-MVWWEZLEFVYHE33KMVRXI2LPNYXTCLRQF4ZUSNBSJAZVGNQ

The BASE32 value encodes the embedder's ident. This is only for uniqueness,
and it does not need to be decoded. This value will also be available from
reading the spec.

Change the ident implementations to delegate to the underlying types.
Decouple (partly) from the `Display` representation so changes there don't
have unintended consequences.

Signed-off-by: J Robert Ray <jrray@jrray.org>