Start allowing embedded package specs to be published
When building a package that contains embedded packages, publish the
embedded package specs too. Update their specs to have the builds
populated with `EmbeddedSource::Ident` set to the parent package.

The act of creating these spec tags may cause runtime errors for older
versions of spk, so this should be put behind a feature flag for enabling
this new behavior later. The code for being able to read these can be
put into place first before enabling this feature.

Signed-off-by: J Robert Ray <jrray@jrray.org>