Get embed stubs to appear in the solve
This is minimally working. When building a package, the "version spec"
for the embedded package needs to exist for the solver to even consider
looking at the package. Then, in `list_package_builds` it usually only
considers tags in the "spk/pkg" tree, but embed stubs don't exist in there,
so also look in the "spk/spec" tree specifically for just embed stubs. It
turns out that the BASE32 values actually do need to be decoded.

The embed stubs are treated as having no components/builds, so the
validation check has to specifically ignore these stubs. The way components
are being handled currently is probably wrong.

In `step_state`, don't allow the solve to finish successfully if the
resolved packages contains a stub that did not get paired up with the
package that provides it.

Signed-off-by: J Robert Ray <jrray@jrray.org>