`EmbeddedPackageValidator` changes for embed stubs
The reasoning here is when validating the embedded packages of a package,
they may not be added to a solve that already contains a non-embedded
version of that package, or if it contains a stub from a different
provider. The solver needs to unwind to before the non-embedded package
was added and add the correct stub instead.

Add test to exercise embedded packages solver solution

Demonstrate that when an embedded package is part of the solution, any
dependencies of the "real" package are discarded and not included.

Closes #346

Signed-off-by: J Robert Ray <jrray@jrray.org>