Refactor `Repository::publish_package`
The logic for creating embedded package stubs should live inside
`publish_package` rather than be the responsibility for any caller of
`publish_package`. Therefore, move the logic out of
`BinaryPackageBuilder::build`. This makes one of the failng tests pass.

This logic for creating stubs is using methods from `Repository` and is not
different for the different types of storage backends. In order to have a
place to have a single implementation of it, a new trait `Storage` was
created to be a new home for lower level methods that act on the storage,
like the original `publish_package` implementations, and then
`Repository::publish_package` can have a default implementation that
delegates to `Storage::publish_package` and also creates the embed stubs.

To avoid ambiguous names, the name was changed to
`Storage::publish_package_to_storage`.

TL;DR: "business logic" for repositories should go in `Repository` and
storage-specific implementations should go in `Storage`.

Signed-off-by: J Robert Ray <jrray@jrray.org>