Clarify source package handling
This introduces a new method on the validator type for validating
recipes in the case of a source build. We also run the validators again
using the package spec that would be generated. This allows the solver
to have more distinct code paths when building from source vs not (in my
opinion). Additionally, this allows us to remove a lot of the special
cases for source packages in the existing validation and solver code and
makes it harder to write code which doesn't account for the case of a
source package being passed.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>