Handle publishing a build individually of a new version
If a user uses `spk publish my-pkg/1.2.3/<build>` this would succeed but
publish only the package and not the "version spec". Then, tools like
`spk ls` would report "Package not found" about that version.

Change publish so that it ensures the "version spec" exists, possibly
overwriting it if the user is doing a force publish.

Signed-off-by: J Robert Ray <jrray@jrray.org>