Fix spfs-monitor failing to read mount namespace
spfs-monitor is launched while the spfs-enter process is owned by root and
is unable to read the mount namespace since it runs as a regular user
(normally). If it were able to read the mount namespace, it might read the
wrong namespace, because spfs-enter hasn't necessarily put the parent
process into the new mount namespace yet.

Have spfs-monitor wait until it reads something from stdin before
attempting to read the mount namespace.

Signed-off-by: J Robert Ray <jrray@imageworks.com>