Push down responsibility of variant options to recipe
Previously, the logic for applying a variant's options along with other
manually specified options was in the cli code for `make-binary` and
`test`.

Push this responsibility down into the builder / tester objects, and inform
them which variant is being built / tested.

The test suite should fail at this point, but it passes!
`Spec::resolve_options` does not do anything with the variant information,
so the wrong options are being used.

Signed-off-by: J Robert Ray <jrray@jrray.org>