Fix variants not producing distinct builds
In the event that a var is mentioned only in the build.variants list, it
will not contribute to the build digest when building that variant. If no
other factors cause the digest to be different, multiple variants will have
the same digest and overwrite each other.

For the example "three-variants.spk.yaml" spec in the new test, these
vars were not contributing to the digest value:

- arch
- distro
- os
- python.abi
- ubuntu

Building that package would perform three builds but only produce one
unique build.

Signed-off-by: J Robert Ray <jrray@jrray.org>