Bake the variant options into build.options
In `Recipe::generate_binary_build`, take the contents of whatever variant
was used and add them to build.options, so the built package of that
variant can be used without knowing what variant it was built from as has
all that variant's requirements listed as requirements of the "default"
variant.

Empty the variant list so the build spec looks like just the single variant
rendered out.

Signed-off-by: J Robert Ray <jrray@jrray.org>