Remove `Error::IO`
Similar to what happened on the spfs side, eliminate automatic translation
of `std::io::Error` into `crate::Error` since this produces poor error
messages without context.

This was inspired by facing an error like this:

```
$ spk test --variant 1 dummy/dummy-e.spk.yaml@build
 INFO Testing dummy/dummy-e.spk.yaml@build...
 INFO Running test #1 variant={dummy-b=3.0.0, os=linux, python.abi=cp37m}
No such file or directory (os error 2)
```

There is little to go on for tracking down where the error is happening.

Now this looks like:
```
$ spk test --variant 1 dummy/dummy-e.spk.yaml@build
 INFO Testing dummy/dummy-e.spk.yaml@build...
 INFO Running test #1 variant={dummy-b=3.0.0, os=linux, python.abi=cp37m}
ERROR Failed to spawn bash process

Caused by:
    No such file or directory (os error 2)
```

It's still not great but has more context clues, enough that I can
recognize that this is probably a case of `Command::current_dir` set to a
path that doesn't exist.

Signed-off-by: J Robert Ray <jrray@jrray.org>