Rescue lost spec test
Port this test that was testing that the local source path was relative to
the spec / recipe / template file. It currently fails.

The file it was moved from is not being imported so this test was not being
run. There is a plan to put some new tests in that file, so it isn't being
removed.

Signed-off-by: J Robert Ray <jrray@jrray.org>