Add test for `Recipe::get_build_requirements` and variants
Verify that a variant can introduce a new package requirement and that
variant entries are properly identified as var or pkg requests.

It works as expected when passing in the fully resolved options for the
correct variant, but this test doesn't prove that the rest of the code will
also call `get_build_requirements` with that.

Signed-off-by: J Robert Ray <jrray@jrray.org>