Make Ident generic over some inner id type
This means that some behaviours can be either enabled or not depending
on the inner type used, with the goal of having some identifiers that
don't have any associated build and others that must have an associated
build. For now, the build is still an Option, with the intention to
change this in the future.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>