Break into many id types for ident
The AnyId, BuildId and VersionId types all now serve different purposes
for what data they can and do contain - specifically in terms of the
availability of the build component. AnyId is still used nearly
everywhere in the codebase to ease the transition, but this lays the
foundation for taking specific Ident types in cases where ambiguity is
not acceptable.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>