Change `clean_render` to clean all users
If deleting an object/payload, any render of that payload will continue to
consume disk space, being a hard link, and also if an object is deleted by
user 1, a render of that object for user 2 would persist but never* be
cleaned, even if user 2 runs `spfs clean`.

Set cap_fowner on spfs-clean so it can delete renders for other users.
The rpm spec file needs this change too, but spk.spec hasn't been updated
to incorporate the merge with spfs yet.

* It would get cleaned if the object had since reappeared and was valid for
deletion again.

Signed-off-by: J Robert Ray <jrray@jrray.org>