Change `spfs clean` to use an object age threshold
This is a safety precaution to prevent accidental false positives for
unattached objects and deleting things that are yet to be attached. Only
objects older than the time before `clean` begins its inventory are
eligible to be deleted.

Objects are typically written to storage before being tagged, so when
cleaning it is possible that a newly written object is discovered before
any tag is created on that object, and then `clean` could consider that
new object unattached.

There could be an arbitrarily long time between writing an object and
tagging it, e.g., pushing a large manifest taking a long time. Use a 15
minute grace period for `clean`.

Signed-off-by: J Robert Ray <jrray@jrray.org>