Handle pipe error when sending "go" to spfs-monitor
Although this notification mechanism could change to just closing the pipe
as the trigger, it is valuable to learn about the pipe error because this
is happening in cases where the spfs-monitor process has exited, and
therefore there is no guarantee about the runtime still existing. Any
objects in use by this new environment are at risk of being cleaned.

Put some context around any errors that happen here so the user feedback is
better.

Signed-off-by: J Robert Ray <jrray@imageworks.com>