Bump rust-cache to v2
Split out the spfs build to a separate step.

Had originally tried to cache these two different flavors of builds
separately with two rust-cache, however when the first one runs it does
something to the container that makes `rustc` not run anymore, making the
second one fail to collect the cache.

Signed-off-by: J Robert Ray <jrray@jrray.org>