Add "embed_stub" methods to `Repository`
Start treating the embed stubs for officially as a different entity type,
in particular so that `MemRepository` can be updated to support embed
stubs.

`Package::Input` doesn't need to exist at this point but it might become
convenient again and was tricky to add so it has been left in place.

`Package::as_recipe` might be controversial but right now the underlying
types allow this to be possible. For embedded stub purposes, this just
needs to clear the build out from the `Package` to become suitable for
publishing as the "version spec".

Signed-off-by: J Robert Ray <jrray@jrray.org>