Fix builds (sometimes) not using bash
Fixes #464. Using setenv in a multi-threaded program is discouraged. I
wasn't able to reproduce the problem in a trivial package but in a complex
package I was able to get the spk build script `echo $SHELL` to print tcsh,
even though it was in fact running the build script with bash. However,
the spfs runtime environment was created with tcsh.

Signed-off-by: J Robert Ray <jrray@jrray.org>