Add concept of "bastion" file for rendering hard links
Hard links share a common owner and mode, so multiple users on the same
host, or different renders of the same blob contents with different
permissions, will compete for changing the current metadata of a hard link.

Introduce another subdirectory into the repo structure:

    spfs-repo/renders/username/work/
    spfs-repo/renders/username/bastion/
    spfs-repo/renders/username/AA/
    spfs-repo/renders/username/AB/
    ...

This new "bastion" directory follows the FSHashStore model, with an
additional subdirectory determined by the mode of the file being rendered.

To render file "foo" with blob digest "ABCD" and mode 0644, the following
steps are taken:

1. Copy (not hard-link) the payload into the bastion location:
   cp spfs-repo/payloads/AB/CD spfs-repo/renders/username/bastion/AB/CD/33188
   # dec 33188 == oct 0100644

2. Hard link the bastion location to the render location:
   ln spfs-repo/renders/username/bastion/AB/CD/33188 spfs-repo/renders/username/AB/CD/foo

3. chmod the render location

Any renders of the same payload that are meant to have a different mode
will use a different bastion file. Multiple renders of the same payload
and mode will all form a hard link group.

Signed-off-by: J Robert Ray <jrray@jrray.org>