Add bastion cleaning to `spfs clean`
Renders are cleaned by wiping out whole manifest directories, which may or
may not cause any single bastion file to become unused. The bastion storage
must be walked to see if there are any unused files in there.

Deleting these files causes a race condition with spfs-render. This could
be addressed by adding a retry loop in spfs-render if the hard link fails
because the source has gone missing.

Signed-off-by: J Robert Ray <jrray@jrray.org>