Adds packages to MustNotAlterExistingFiles build error messages
This generates a custom error message based on the file that was altered
that shouldn't have been. The package, or packages, that contain the file
and the exact differences detected are added to the message. This changed
the schema validators to return Errors containing data instead of optional
Strings, and passing the seen and conflicting_packages data down into the
validation levels from the build processes.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>