Add a build-time warning about shadowed files
When multiple layers of a build environment provide the same file, emit a
warning. This can be from packages that have captured files they shouldn't
have, an artifact of ignorning diff errors due to mode changes, pre-spfs
render proxy files (a problem specific to SPI). Otherwise, it could be
useful to know that a package's dependencies happen to have the same files
in them.

Signed-off-by: J Robert Ray <jrray@imageworks.com>