Add filter mode for components file filtering
In order to make it easier to divide files up into different components,
add a filter mode to enable "Exclusive" filtering, so a file is only added
to a component if it hasn't already been matched by an earlier component.

In this example pulled from the documentation:

    install:
      components:
      - name: run
        # only the compiled libraries are needed at runtime
        files: [lib/mylib*.so]
      - name: build
        # but everything else (debug symbols or static libraries, for example)
        # should be pulled in when building against this package
        files: ['*']
        filter_mode: Exclusive

As the "else" in "everything else" implies, the `build` component will
include all the files that were not previously matched by the `run`
component. Specifically, it will not contain any files matching the
pattern `lib/mylib*so`.

The default filter mode retains the original behavior.

Signed-off-by: J Robert Ray <jrray@jrray.org>