Add a `debug` cli command
This new command is useful for bringing in all the `src` package components
for the packages found in the current environment. Whatever packages that
are found in the current runtime are mapped to the `src` build and those
layers are added to the runtime layer stack. Then, in theory, a debugger
will be able to find the sources during a debugging session.

Signed-off-by: J Robert Ray <jrray@imageworks.com>