Fix spurious failure to render
When multiple processes are simultaneously trying to render the same
manifests, as in doing `spk env xyz` in multiple processes, and the spfs
blobs have not already been rendered, there is a race condition where one
process fails to create the hard link from the proxy file, because the
other process is renaming on top of the same proxy file. Even though the
rename should be an atomic operation, it is observed to fail regularly
when experimenting with running 20 copies of the same `spfs run` command.

Curiously, even the 3 retries doesn't resolve the issue, although some
processes are observed succeeding after 1 or 2 retries. Adding a sleep
between retries will make the error go away.

Switching to `persist_noclobber` works consistently, because it does not
clobber the existing file, if any. This was the intended semantics all
along, with this code ignoring any `AlreadyExists` error.

Signed-off-by: J Robert Ray <jrray@imageworks.com>