Remove RecipeOps trait and replace with Satisfy
This trait was ambiguous in that is was not a recipe but just a small
collection of recipe-ish functions. The to_ident function was easily
moved to the recipe trait, allowing a removal of the redundant Ident
associated type, and the rest of the trait was providing function for
checking request satisfaction, which is replaced by a new Satisfy trait
with clearer responsibilities.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>