Make Ident a generic type
The pattern is <base>/<target> which allows for defining different
spcificity levels as well as nice composition of different ident types.
This initial commit defines different ident types as well as an AnyIdent
with an optional build that replaces the old ident nearly everywhere.
Next steps are to use more specific identifier types where applicable

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>