Update recipe and package to have different idents
This mean that the v0 spec type needed to be updated to be generic over
the internal ident type, as well as the storage interfaces using the
appropriate ident types for loading recipes vs packages.

This was largely an exercise in refactoring and separating logic that
used to be conditional on the presence of a build into distict functions
instead.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>