Clean up ident impls and relations
Uses macros to share common 'inheritance' instead of having a deref
implementation. This works better accross types where dereferencing to
the base is not as useful as the target (eg located idents) but also
creates a safer usage where ident types are not converted silently,
instead needing an explicit conversion call by the developer.

Additionally, add a HasLocation trait to complete the set.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>