Set sync policy when syncing tag in spk-launcher
The default policy will not pick up changes to an existing tag.

spk-launcher will resolve a tag on the remote to find its manifest digest,
and then use that digest for the name of directory used to render out the
manifest. But when it renders it resolves the local tag which might already
exist but be older content. The `LatestTags` policy is needed to ensure the
local content is the same as the remote content (subject to a race with the
tag being modified between resolving it and syncing it).

Signed-off-by: J Robert Ray <jrray@imageworks.com>