Adds tasks and threading to impossible checks, and cleanup updates
This extends the impossible request checks use of tasks and threads to
reduce the time overhead the checks add to a solve. The
ImpossibleRequestChecker is made thread-safe by using DashMap and
Atomic counters.

Internally, the ImpossibleRequestChecker now uses tasks, threads, and
a message channels to check a build for impossible requests. The
requirements from the build are still checked sequentially, but for
each one it sets up a channel and launches one task, the maker task,
that then generates a task per version. The version tasks read and
check all the builds for that version, and send a message on the
channel as soon as they find a either valid build, or no valid
builds. The tasks are all spawned and run in parallel on other
threads, and they are stopped as soon as a valid build is found.

The two of the three place where the Solver runs impossible request
checks have also been changed to use tasks:

- The initial request checks set up a task per request and runs them
  concurrently in the current thread, via async. Each makes a single
  call to the impossible request checker (which will launch more tasks
  in parallel behind the scenes as described above).

- The build keys checks set up a task per build and runs them in
  parallel in separate threads. Each makes a single call to the
  impossible request checker (which will launch more tasks as above).

- The pre-resolve build validation check doesn't make any additional tasks.

Based on our current testing, that combination of tasks/threads seemed
to perform the best compared with running the impossible checks sequentially.
The slowest part of the process is reading the package specifications and
the package's list of build component names from disk. The checks speed up
dramatically once the all data is cached.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>