Adjust warning/error behavior of `spk rm`
If the spec was not found but the package was removed, it would first
warn that the spec was not found and then return an error because it was
not found.

First, don't print a warning if this is also going to error for the same
cause.
Second, don't error if only the spec or only the package was not found.

This makes the experience better if doing `spk rm a b c` and there were
"warnings" with package `a`; it will continue working to delete `b` and `c`
rather than stop after partially deleting `a`.

Signed-off-by: J Robert Ray <jrray@imageworks.com>