Formalize embedded packages as a generic type
Previously, embedded packages were assumed to be the enum spec type,
which created an undesirable loop in the type hierearchy. Now, packages
define their embedded stub type and it can be separate. This also
extends to component specs, which also now are generic over the type of
embedded package that they allow.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>