Remove use of v0::Opt from binary builder
This changes the current inheritance mechanism from being calculated
based on the specific v0::Opt::inheritance field, instead expecting the
package itself to identify downstream requests that need to be present.
In short, this means that users need to add explicit requests to
satisfy the upstream packages. As much as this is more work for users,
it ensures that the final nature of each one remains up to the package
definition so that they can be based on when conditions and/or attached
only to specific components which is not possible now.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>