Add filter concept for when collecting files after build
This is to address an issue with packages collecting extraneous files that
were not intended to be included in a build.

A build script may do unusual things like `touch` or temporarily rename
files that belong to one of its build dependencies. For example:

    pkg-dep:
        - include/dep/a.h
        - include/dep/b.h
        - include/dep/c.h

Let's say package my-pkg depends on pkg-dep at build time, and has this
as part of its build script:

    - mv /spfs/include/dep /spfs/include/.dep.hide
    - ...
    - mv /spfs/include/.dep.hide /spfs/include/dep

On the "validating package contents" phase, nothing here is amiss, by the
end of the build script execution, none of the existing files have been
modified.

On the "committing package contents" phase, the upperdir is walked and all
the contents of `include/dep` will be discovered in the upperdir due to the
activities of the build script. As a result, pkg-dep ends up with a copy
of `include/dep/*.h` inside of it.

In this commit, the "diff" calculated in the "validating package contents"
is saved, because this is a more accurate representation (from spk's point
of view) of what the build script _actually_ changed. Then, the lower level
spfs code now optionally takes this diff as an include filter. When walking
the upperdir, the entries are only captured if they are also found in the
include filter.

Back to the example above, now pkg-dep will not contain `/include/dep/*.h`,
because all of those files are tagged as "unchanged" in the diff, and will
be passed over during the second phase.

Signed-off-by: J Robert Ray <jrray@imageworks.com>