Rework how solver deals with embedded packages
The proliferation of embedded packages has an adverse effect of increasing
the search space and making it a common problem that requests involving
embedded packages cannot find a solution in a resonable amount of time.
There isn't a good way to know if any given embedded stub should be
preferred or avoided and making the wrong choice can lead down the wrong
path. It is also wrong to prefer or avoid all embedded stubs categorically,
since any given set of requests may need some embedded packages but not
others.

The approach taken here is to avoid considering any embedded packages until
a package is needed for the solution that is a parent of embedded packages.

First of all, no embedded stubs are ever considered during the solve. Note
that embedded stubs still provide value to the solver by letting it make
decisions based on the knowledge of what packages do or don't exist (even
embedded ones).

When a parent of an embedded package is needed, but any of its embedded
packages already exist as part of the solution in progress, a `StepBack`
change is taken that jumps to the state immediately before one of those
packages was added to the state. It also prioritizes the embedded package
parent as the next package to attempt to add to the solution (again).

For example, if "boost" is embedded by "boost-python", the solver might
take this route:

    > RESOLVE pkg-a
    >> RESOLVE pkg-b
    >>> RESOLVE boost
    >>>> RESOLVE pkg-c
    >>>>> TRY boost-python
    >>>>> BLOCKED Package boost-python embeds package already resolved: boost
    >>> RESOLVE boost-python
    >>> RESOLVE boost:embedded
    >>>> RESOLVE pkg-c

After BLOCKED it jumps back to after pkg-b was resolved but before boost.
It solves for boost-python while in that state and resumes trying to solve
for the remaining requests.

Signed-off-by: J Robert Ray <jrray@imageworks.com>